refactor: split newton_backend.py into package modules
- Split 2,159-line newton_backend.py monolith into 13 focused modules
- _types.py: NewtonConfig, solver/render constants, procedural robot registry
- _registry.py: lazy Newton/Warp imports, model path resolution, procedural builder
- _core.py: thin NewtonBackend orchestrator delegating to submodules
- _scene.py: create_world, replicate, reset, step, finalize_model
- _robots.py: add_robot with URDF/MJCF/USD/procedural fallback
- _objects.py: cloth, cable, and particle management
- _rendering.py: render and get_observation
- _recording.py: video recording
- _policy.py: policy execution and mock policy
- _diffsim.py: differentiable simulation via wp.Tape
- _sensors.py: contact, IMU, and tiled camera sensors
- _ik.py: Jacobian-based inverse kinematics
- Renamed newton_gym_env.py to _gym_env.py for internal module naming consistency
- Updated imports in envs.py, factory.py, rl_trainer.py, training/evaluate.py
- All public interfaces unchanged
